Move utils functions into lib folder
