+++ lib/streamlit/components/v1/custom_component.py
@@ -20,7 +20,7 @@ from typing import TYPE_CHECKING, Any
 from streamlit import _main, type_util
 from streamlit.components.types.base_custom_component import BaseCustomComponent
 from streamlit.elements.form import current_form_id
-from streamlit.elements.policies import check_cache_replay_rules
+from streamlit.elements.lib.policies import check_cache_replay_rules
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.Components_pb2 import ArrowTable as ArrowTableProto
 from streamlit.proto.Components_pb2 import SpecialArg
