+++ lib/streamlit/elements/widgets/color_picker.py
@@ -21,12 +21,12 @@ from typing import cast
 
 import streamlit
 from streamlit.elements.form import current_form_id
-from streamlit.elements.policies import (
+from streamlit.elements.lib.policies import (
     check_cache_replay_rules,
     check_callback_rules,
     check_session_state_rules,
 )
-from streamlit.elements.utils import get_label_visibility_proto_value
+from streamlit.elements.lib.utils import get_label_visibility_proto_value
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.ColorPicker_pb2 import ColorPicker as ColorPickerProto
 from streamlit.runtime.metrics_util import gather_metrics
