+++ lib/streamlit/elements/widgets/data_editor.py
@@ -55,6 +55,11 @@ from streamlit.elements.lib.column_config_utils import (
     update_column_config,
 )
 from streamlit.elements.lib.pandas_styler_utils import marshall_styler
+from streamlit.elements.lib.policies import (
+    check_cache_replay_rules,
+    check_callback_rules,
+    check_session_state_rules,
+)
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.Arrow_pb2 import Arrow as ArrowProto
 from streamlit.runtime.metrics_util import gather_metrics
@@ -787,11 +792,6 @@ class DataEditorMixin:
         import pyarrow as pa
 
         # Import here to avoid cyclic import warning
-        from streamlit.elements.policies import (
-            check_cache_replay_rules,
-            check_callback_rules,
-            check_session_state_rules,
-        )
 
         key = to_key(key)
