+++ lib/streamlit/elements/widgets/radio.py
@@ -19,12 +19,15 @@ from textwrap import dedent
 from typing import TYPE_CHECKING, Any, Callable, Generic, Sequence, cast
 
 from streamlit.elements.form import current_form_id
-from streamlit.elements.policies import (
+from streamlit.elements.lib.policies import (
     check_cache_replay_rules,
     check_callback_rules,
     check_session_state_rules,
 )
-from streamlit.elements.utils import get_label_visibility_proto_value, maybe_coerce_enum
+from streamlit.elements.lib.utils import (
+    get_label_visibility_proto_value,
+    maybe_coerce_enum,
+)
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.Radio_pb2 import Radio as RadioProto
 from streamlit.runtime.metrics_util import gather_metrics
